refactor(web): move ConnectionForm files to dedicated subdirectory
Organizes all ConnectionForm-related components into connection-form/
subdirectory to separate form logic from top-level network components
(ConnectionsTable, NetworkPage, etc.).
