+++ web/src/components/network/connection-form/BindingModeSelector.test.tsx
@@ -24,7 +24,7 @@ import React from "react";
 import { screen } from "@testing-library/react";
 import { installerRender } from "~/test-utils";
 import { useAppForm } from "~/hooks/form";
-import { connectionFormOptions } from "~/components/network/ConnectionForm";
+import { connectionFormOptions } from "./ConnectionForm";
 import BindingModeSelector from "./BindingModeSelector";
 
 function TestForm() {
