+++ web/src/components/network/connection-form/BridgeSettings.test.tsx
@@ -24,7 +24,7 @@ import React from "react";
 import { screen } from "@testing-library/react";
 import { installerRender } from "~/test-utils";
 import { useAppForm } from "~/hooks/form";
-import { connectionFormOptions, BridgeStpMode } from "~/components/network/ConnectionForm";
+import { connectionFormOptions, BridgeStpMode } from "./ConnectionForm";
 import { DeviceState } from "~/types/network";
 import { CONNECTION_TYPE } from "~/utils/network";
 import BridgeSettings from "./BridgeSettings";
