+++ web/src/components/network/connection-form/DeviceSelector.test.tsx
@@ -26,7 +26,7 @@ import { installerRender } from "~/test-utils";
 import { useAppForm } from "~/hooks/form";
 import { DeviceState } from "~/types/network";
 import { CONNECTION_TYPE } from "~/utils/network";
-import { connectionFormOptions } from "~/components/network/ConnectionForm";
+import { connectionFormOptions } from "./ConnectionForm";
 import DeviceSelector from "./DeviceSelector";
 
 const mockDevices = [
