+++ web/src/components/network/connection-form/IpSettings.test.tsx
@@ -24,7 +24,7 @@ import React from "react";
 import { screen } from "@testing-library/react";
 import { installerRender } from "~/test-utils";
 import { useAppForm } from "~/hooks/form";
-import { connectionFormOptions, FormIpMode } from "~/components/network/ConnectionForm";
+import { connectionFormOptions, FormIpMode } from "./ConnectionForm";
 import IpSettings from "./IpSettings";
 
 function TestForm({
