+++ web/src/components/network/connection-form/connectionFormValidation.ts
